Migrate serverEnv to utils
