--- src/constants/system.ts
@@ -1,6 +1,0 @@
-export enum serverEnv {
-  development = "development",
-  test = "test",
-  staging = "staging",
-  production = "production",
-}
